refactor app config
